+++ src/formats/mnx/CMakeLists.txt
@@ -2,6 +2,7 @@ set(DENIGMA_FORMAT_MNX_SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/mnx.cpp
     ${CMAKE_CURRENT_LIST_DIR}/mnx_converter.cpp
     ${CMAKE_CURRENT_LIST_DIR}/mnx_enums.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/mnx_expressions.cpp
     ${CMAKE_CURRENT_LIST_DIR}/mnx_formatted_text.cpp
     ${CMAKE_CURRENT_LIST_DIR}/mnx_global.cpp
     ${CMAKE_CURRENT_LIST_DIR}/mnx_layouts.cpp
